refactored code to allow callable for compatibility with gtest
