+++ isis/src/lro/apps/lronaccal/lronaccal.h
@@ -1,0 +1,13 @@
+#ifndef lronaccal_h
+#define lronaccal_h
+
+#include "Cube.h"
+#include "lronaccal.h"
+#include "UserInterface.h"
+
+namespace Isis{
+  extern void lronaccal(UserInterface &ui);
+  extern void lronaccal(Cube *iCube, UserInterface &ui);
+}
+
+#endif
